internal: Move actions to action crate. Rename labels. (#187)
* Move actions to action crate.

* Move node to own file.

* Rename labels.
