+++ crates/action-runner/Cargo.toml
@@ -5,23 +5,17 @@ edition = "2021"
 
 [dependencies]
 moon_action = { path = "../action" }
-moon_cache = { path = "../cache" }
 moon_config = { path = "../config" }
 moon_error = { path = "../error" }
-moon_hasher = { path = "../hasher" }
 moon_lang = { path = "../lang" }
-moon_lang_node = { path = "../lang-node" }
 moon_logger = { path = "../logger" }
 moon_project = { path = "../project" }
-moon_terminal = { path = "../terminal" }
-moon_toolchain = { path = "../toolchain" }
-moon_utils = { path = "../utils" }
-moon_vcs = { path = "../vcs" }
 moon_workspace = { path = "../workspace" }
-clap = { version = "3.2.8", features = ["derive"] }
 petgraph = "0.6.2"
 thiserror = "1.0.31"
 tokio = { version = "1.19.2", features = ["full"] }
 
 [dev-dependencies]
+moon_cache = { path = "../cache" }
+moon_utils = { path = "../utils" }
 insta = "1.15.0"
