--- crates/action-runner/src/actions/mod.rs
@@ -1,10 +1,0 @@
-mod install_node_deps;
-mod run_target;
-mod setup_toolchain;
-mod sync_node_project;
-mod target;
-
-pub use install_node_deps::*;
-pub use run_target::*;
-pub use setup_toolchain::*;
-pub use sync_node_project::*;
