+++ crates/action-runner/src/errors.rs
@@ -1,7 +1,5 @@
 use moon_error::MoonError;
 use moon_project::ProjectError;
-use moon_toolchain::ToolchainError;
-use moon_vcs::VcsError;
 use moon_workspace::WorkspaceError;
 use thiserror::Error;
 
@@ -18,12 +16,6 @@ pub enum ActionRunnerError {
 
     #[error(transparent)]
     Project(#[from] ProjectError),
-
-    #[error(transparent)]
-    Toolchain(#[from] ToolchainError),
-
-    #[error(transparent)]
-    Vcs(#[from] VcsError),
 
     #[error(transparent)]
     Workspace(#[from] WorkspaceError),
