+++ crates/action-runner/src/lib.rs
@@ -1,10 +1,9 @@
-mod actions;
-mod context;
 mod dep_graph;
 mod errors;
+mod node;
 mod runner;
 
-pub use context::*;
 pub use dep_graph::*;
 pub use errors::*;
+pub use node::Node;
 pub use runner::*;
