+++ crates/action/Cargo.toml
@@ -2,3 +2,19 @@
 name = "moon_action"
 version = "0.1.0"
 edition = "2021"
+
+[dependencies]
+moon_cache = { path = "../cache" }
+moon_config = { path = "../config" }
+moon_error = { path = "../error" }
+moon_hasher = { path = "../hasher" }
+moon_logger = { path = "../logger" }
+moon_project = { path = "../project" }
+moon_terminal = { path = "../terminal" }
+moon_toolchain = { path = "../toolchain" }
+moon_utils = { path = "../utils" }
+moon_vcs = { path = "../vcs" }
+moon_workspace = { path = "../workspace" }
+clap = { version = "3.2.8", features = ["derive"] }
+thiserror = "1.0.31"
+tokio = { version = "1.19.2", features = ["full"] }
