+++ crates/action/src/lib.rs
@@ -1,3 +1,16 @@
 mod action;
+mod context;
+mod errors;
+mod install_node_deps;
+mod run_target;
+mod setup_toolchain;
+mod sync_node_project;
+mod target;
 
 pub use action::*;
+pub use context::*;
+pub use errors::ActionError;
+pub use install_node_deps::*;
+pub use run_target::*;
+pub use setup_toolchain::*;
+pub use sync_node_project::*;
